requirements.txt
matplotlib==2.1.1
nibabel==3.2.0
numpy==1.16.4
PyYAML==5.4.1
torch==1.5.1+cpu
tqdm==4.54.1
